backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfminer.six==20221105
python-docx==1.0.1
openai==1.3.0
python-dotenv==1.0.0
supabase==2.0.3
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 